build(assemblyscript): Removed assemblyscript
This commit removes both the assemblyscript dev dependency
and the scaffold assemblyscript project created with it.

Doing some benchark I found out that the current implementation
in plain Typescript/Javascript is performant enough to sustain on its
own the simulation of a cellular automaton with 15**3 cells.
This means that the complexities of managing WebAssembly and the
exchange of data betweeen WASM and JS are  not needed.

I suspect that theese "unexpected" performances have to do with the
automatic JIT optimization available in V8 as well as SpiderMonkey, more
test willl follow to confirm this hypothesis.