fix(Automaton): Fixed a rendering bug
The bug was caused by the computation of the ageMatrixIdx starting
from the given cartesian coordinates, this means that most of the
ageMatrixIdx computed were invalid or wrong (negative, multiple zero, ...).
The current fix is to decouple the generation of the linearized index and of
the coords from some simple counter x,y,z and pass all as props.