fix(core): Fixed a logic bug
The bug was caused by the  missing reset of the matrixIdx variable
during the execution of the init() function.
The missed reassignment to the start value meant that sometimes matrixB
(zero initialized) was used as oldGen during the first call of nextGeneration()
returning a dead automaton almost immediately after the initialization.