fix(config): Fixed next.config.js to account for production env
The current commit adds a "dynamic" next.config.js file based on the current env.
Since the site is deployed to GitHub Pages as a project page at the following url:
{username}.github.io/{projectname} but Next.js expects  the assetts to be served
at the root location: {username}.github.io.
Since this is not possible in GitHub Pages, this commit fixes the issue by hardcoding
the basePath and assetPath variable in the configuration file.

Credits: https://wallis.dev/blog/deploying-a-next-js-app-to-github-pages